+++ src/Handler.php
@@ -16,12 +16,17 @@ final class Handler
     private bool $wasAsync;
     private bool $resetted = false;
 
-    public function __construct()
+    private function __construct()
     {
         /** @var Map<Signal, Sequence<callable(Signal, Info): void>> */
         $this->handlers = Map::of();
         $this->wasAsync = \pcntl_async_signals();
         \pcntl_async_signals(true);
+    }
+
+    public static function main(): self
+    {
+        return new self;
     }
 
     /**
@@ -46,23 +51,23 @@ final class Handler
     public function remove(callable $listener): void
     {
         $handlers = $this->handlers->map(
-            static fn($_, Sequence $listeners): Sequence => $listeners->filter(
-                static fn(callable $callable): bool => $callable !== $listener,
+            static fn($_, $listeners) => $listeners->exclude(
+                static fn($callable) => $callable === $listener,
             ),
         );
-        $_ = $handlers->foreach(static function(Signal $signal, Sequence $listeners): void {
+        $_ = $handlers->foreach(static function($signal, $listeners): void {
             if ($listeners->empty()) {
                 \pcntl_signal($signal->toInt(), \SIG_DFL); // restore default handler
             }
         });
-        $this->handlers = $handlers->filter(
-            static fn($_, Sequence $listeners): bool => !$listeners->empty(),
+        $this->handlers = $handlers->exclude(
+            static fn($_, $listeners) => $listeners->empty(),
         );
     }
 
     public function reset(): void
     {
-        $_ = $this->handlers->foreach(static function(Signal $signal): void {
+        $_ = $this->handlers->foreach(static function($signal): void {
             \pcntl_signal($signal->toInt(), \SIG_DFL);
         });
         \pcntl_async_signals($this->wasAsync);
@@ -78,39 +83,34 @@ final class Handler
         return $this
             ->handlers
             ->get($signal)
-            ->match(
-                static fn($listeners) => $listeners,
-                function() use ($signal) {
-                    \pcntl_signal($signal->toInt(), function(int $signo, mixed $siginfo): void {
-                        $this->dispatch(Signal::of($signo), $siginfo);
-                    });
+            ->otherwise(function() use ($signal) {
+                \pcntl_signal($signal->toInt(), function($signo, $siginfo): void {
+                    $this->dispatch(Signal::of($signo), $siginfo);
+                });
 
-                    /** @var Sequence<callable(Signal, Info): void> */
-                    return Sequence::of();
-                },
-            );
+                /** @var Maybe<Sequence<callable(Signal, Info): void>> */
+                return Maybe::nothing();
+            })
+            ->toSequence()
+            ->flatMap(static fn($listeners) => $listeners);
     }
 
     private function dispatch(Signal $signal, mixed $info): void
     {
         $info = \is_array($info) ? $info : [];
-        /** @psalm-suppress MixedArgument */
-        $structure = new Info(
-            Maybe::of($info['code'] ?? null)->map(static fn($code) => new Signal\Code($code)),
-            Maybe::of($info['errno'] ?? null)->map(static fn($errno) => new Signal\ErrorNumber($errno)),
-            Maybe::of($info['pid'] ?? null)->map(static fn($pid) => new Signal\SendingProcessId($pid)),
-            Maybe::of($info['uid'] ?? null)->map(static fn($uid) => new Signal\SendingProcessUserId($uid)),
-            Maybe::of($info['status'] ?? null)->map(static fn($status) => new Signal\Status($status)),
+        $structure = Info::of(
+            Maybe::of($info['code'] ?? null)->map(Signal\Code::of(...)),
+            Maybe::of($info['errno'] ?? null)->map(Signal\ErrorNumber::of(...)),
+            Maybe::of($info['pid'] ?? null)->map(Signal\SendingProcessId::of(...)),
+            Maybe::of($info['uid'] ?? null)->map(Signal\SendingProcessUserId::of(...)),
+            Maybe::of($info['status'] ?? null)->map(Signal\Status::of(...)),
         );
 
-        /** @var Sequence<callable(Signal, Info): void> */
-        $listeners = $this
+        $_ = $this
             ->handlers
             ->get($signal)
-            ->match(
-                static fn($listeners) => $listeners,
-                static fn() => Sequence::of(),
-            );
-        $_ = $listeners->foreach(static fn(callable $listener) => $listener($signal, $structure));
+            ->toSequence()
+            ->flatMap(static fn($listeners) => $listeners)
+            ->foreach(static fn($listener) => $listener($signal, $structure));
     }
 }